@@ -7,19 +7,20 @@
   "scripts": {
     "build": "tsc",
     "watch": "tsc -w",
-    "cdk": "cdk"
+    "cdk": "cdk synth -c oauthToken=$OAUTHTOKEN"
   },
   "devDependencies": {
     "@types/node": "8.10.45",
-    "typescript": "^3.3.3333",
+    "aws-cdk": "^1.1.0",
     "ts-node": "^8.1.0",
-    "aws-cdk": "^0.36.2"
+    "typescript": "^3.5.3"
   },
   "dependencies": {
-    "@aws-cdk/aws-codebuild": "^0.36.2",
-    "@aws-cdk/aws-codepipeline": "^0.36.2",
-    "@aws-cdk/aws-codepipeline-actions": "^0.36.2",
-    "@aws-cdk/core": "^0.36.2",
+    "@aws-cdk/aws-codebuild": "^1.1.0",
+    "@aws-cdk/aws-codepipeline": "^1.1.0",
+    "@aws-cdk/aws-codepipeline-actions": "^1.1.0",
+    "@aws-cdk/aws-secretsmanager": "^1.1.0",
+    "@aws-cdk/core": "^1.1.0",
     "source-map-support": "^0.5.9"
   }
 }
