@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 import 'source-map-support/register';
-import cdk = require('@aws-cdk/core');
 import { CicdStack } from '../lib/cicd-stack';
+import { App } from '@aws-cdk/core';
 
-const app = new cdk.App();
+const app = new App();
 new CicdStack(app, 'CicdStack');
